feat(remix)!: Remove import-in-the-middle instrumentation
Migrate `@sentry/remix` server instrumentation fully to the channel-based
path and drop the OpenTelemetry `RemixInstrumentation` module patcher, so the
package no longer relies on the import-in-the-middle ESM loader hook.
