+++ packages/remix/package.json
@@ -57,8 +57,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@opentelemetry/api": "^1.9.1",
-    "@opentelemetry/instrumentation": "^0.220.0",
     "@remix-run/router": "^1.23.3",
     "@sentry/cli": "^2.58.6",
     "@sentry/conventions": "^0.16.0",
