+++ packages/remix/src/server/integrations/RemixIntegration.ts
@@ -1,8 +1,6 @@
 import type { IntegrationFn } from '@sentry/core';
 import { defineIntegration, getClient } from '@sentry/core';
-import { isOrchestrionInjected } from '@sentry/server-utils/orchestrion';
 import { instrumentRemix } from './tracing-channel';
-import { addRemixSpanAttributes, instrumentRemixWithOpenTelemetry } from './opentelemetry';
 import type { RemixOptions } from '../../utils/remixOptions';
 
 const INTEGRATION_NAME = 'Remix' as const;
@@ -17,18 +15,7 @@ const _remixIntegration = (() => {
         ? options?.captureActionFormDataKeys
         : undefined;
 
-      if (isOrchestrionInjected()) {
-        instrumentRemix(actionFormDataAttributes);
-      } else {
-        instrumentRemixWithOpenTelemetry({ actionFormDataAttributes });
-      }
-    },
-    setup(client) {
-      if (!isOrchestrionInjected()) {
-        client.on('spanStart', span => {
-          addRemixSpanAttributes(span);
-        });
-      }
+      instrumentRemix(actionFormDataAttributes);
     },
   };
 }) satisfies IntegrationFn;
